q - 4/requirements.txt
fastapi
grpcio
grpcio-tools
protobuf
uvicorn[standard]
websockets
aiofiles
